添加服务器间传输剩余数据的处理(windows)

+++ purenessscopeserver/purenessscopeserver/PurenessScopeServer/Reactor/ConnectClient.cpp
@@ -363,14 +363,14 @@ int CConnectClient::RecvData()
     if (m_pCurrMessage->length() > 0)
     {
         //将缓冲没处理的数据，放入下次缓冲
-        ACE_Message_Block* pmbSave = App_MessageBlockManager::instance()->Create(m_pCurrMessage->length());
+        ACE_Message_Block* pmbSave = App_MessageBlockManager::instance()->Create((uint32)m_pCurrMessage->length());
 
         if (NULL != pmbSave)
         {
-            memcpy_safe(pmbSave->wr_ptr(), m_pCurrMessage->length(), m_pCurrMessage->rd_ptr(), m_pCurrMessage->length());
+            memcpy_safe(pmbSave->wr_ptr(), (uint32)m_pCurrMessage->length(), m_pCurrMessage->rd_ptr(), (uint32)m_pCurrMessage->length());
             pmbSave->wr_ptr(m_pCurrMessage->length());
             m_pCurrMessage->reset();
-            memcpy_safe(m_pCurrMessage->wr_ptr(), pmbSave->length(), pmbSave->rd_ptr(), pmbSave->length());
+            memcpy_safe(m_pCurrMessage->wr_ptr(), (uint32)pmbSave->length(), pmbSave->rd_ptr(), (uint32)pmbSave->length());
             m_pCurrMessage->wr_ptr(pmbSave->length());
             App_MessageBlockManager::instance()->Close(pmbSave);
         }
